Fix guide-key breakage
1958678754 (0.26.0) introduced a customisable variable
`dotspacemacs-leader-key`. Hardcoded instances of the leader were
replaced by references to this variable. An error in this refactoring
has broken guide-key.

This change fixes a bug where the variable name was added to a quoted
list, inserting the symbol and not the variable's value.

After this change, guide-key should display again.