Add keys to copy & paste full buffer
When you are typing in a browser and then you realize you really want
Emacs to compose this piece of text, then you would want to copy from
browser to an Emacs buffer, edit it inside Emacs and finally copy to
clipboard and paste it into the original application. These keys help
you to do exactly that.