Improve cider-test integration
* `mtt` now reloads test ns before running all tests
* added `mtT` to reload test ns and run focused test
* added `mtr` to reload test ns and re-run failed tests for namespace

The test namespace reloading is a workflow optimization. The alternative
is a minimum of two commands (`SPC m e b` or `SPC m e f` and `SPC m t
t`) in order to re-run a single test. When driving development from
tests, this is onerous at best (and I couldn't think of a good reason
_not_ to reload the test ns).