Add Spacemacs Snippets
There are many stuff that we have in spacemacs which are creating new
layers and also creating micro-states. This Snippets try to help with
the process (the micro-state is a work in progress).

I also set the binding `C-s-/` to `(yas-next-field-or-maybe-expand)`
because `tab` doesn't have the behaviour I quite want (to be able to
expand a snippet like (defun) inside the new-package snippet and still
be able to continue in the outer snippet afterwards. With `tab` this
is *not* possible, but with `C-M-/` it *is*.

Changed variable name

Changed keybinding

Renamed snip-dir to spacemacs-snippets-dir

Moved `setq` outsidde of inizialization function, and the keybindning
too